Add main application structure and templates
This commit introduces the main structure of the application, including the introduction of the main app configuration, rendering repository, templates, and handlers repository. It also adds base, home, and about templates used for the web pages in HTML. It sets up the home and about routes and listens on a specified port to serve the application. Additionally, it includes several new files and directories in the .gitignore file, specifically for JetBrains IDE and Go templates.